chore(package): remove unnecessary binstubs path
NPM (and Yarn) both add the relevant binstubs path to `$PATH`, so you don't need to add it explicitly. On top of that, this hardcodes a `node_modules` layout that may not always be correct. If you were to use Yarn v2 with Zero Install, for example, it never unpacks packages into `node_modules` at all. NPM may do something similar in the future.