locator: attempt to fix Safari RangeError exceptions
Hopefully circumvent an issue in Safari 13+ on both Mac and iOS (also including
iOS Chrome and other Safari iOS derivatives). Safari was very occasionally and
apparently not deterministically throwing a "RangeError: Array size is not a
small enough positive integer." exception seemingly within the second .map of
the original code which was now converted into a for-loop. This second .map
contained a nested .map call over the same array instance which was the chained
result from previous calls to .map, .filter and .sort which potentially caused
this bug in Safari?
Also see https://github.com/cozmo/jsQR/issues/157 and
https://bugs.webkit.org/show_bug.cgi?id=211619#c3